# Simplify `dropdims(::Transpose)` and `insertdims(::PermutedDimsArray)` etc.