@mcabbott 2025-01-06
Pre-0.12 bump?

After #56637 this would need to be split into two, but I won't bother if nobody will review it.