@mcabbott 2025-01-13
No, only half of this PR is LinearAlgebra, the other half is Base. It can stay open for discussion?