@adienes 2025-09-03
just spitballing: the fact that both `dropdims` and `insertdims` forward to `reshape` suggest that maybe `reshape(::PermutedDimsArray)` should be specialized rather than each of `dropdims` and `insertdims`. also `reshape` is not required to return an array of the same wrapper type as its input, so maybe we could get away with actually just returning a reshape of a view of the parent